Remove code generated whitelisting, generate to separate files
- Revert "[pytorch] move force schema registration output into a separate file (#36284)"
  This reverts commit 126d00c8ddcca996a5045f6a6c59cf86c120936c.
- Revert "[pytorch] force c10 schema registration for custom build"
  This reverts commit 61b680c01250c420603ae2937a2f2f16adebc475.
- Revert "[pytorch] aten codegen to filter backends for default mobile build"
  This reverts commit 6792bafa7258e66cf66f738fac6ab8316c4bc0f8
- Revert "[pytorch] codegen flags to whitelist op registrations / generate to separate files (#32451)"
  This reverts commit 69283388caa1fabe695ca0209ea03bed830de9be.

The new plan on record is that we will control whether or not
we should register an operator or not by a non-codegenned mechanism
(constexpr conditional).

This PR doesn't actually implement this new mechanism yet (for ease
of review), which means it will break all of our selective builds.
I intend to squash it with a future PR later.

ghstack-source-id: 3c9b284fec5f276d14671a0c9e2a80cb03d53e36
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39350